updating to use any engine

+++ main.js
@@ -41,7 +41,8 @@ exports.startListening = function(timeout) {
 }
 
 exports.stopListening = function() {
-  child.kill();
+  if (child != null)
+    child.kill();
 }
 
 
@@ -67,7 +68,8 @@ process.once("uncaughtException", function(error) {
   // Our assumption here is that any other code listening for an uncaught
   // exception is going to do the sensible thing and call process.exit().
   if (process.listeners("uncaughtException").length === 0) {
-    child.kill();
+    if (child != null)
+      child.kill();
     throw error;
   }
 });
